[r2dbe_vdif, r2dbe_top] Re-done such that timing is clear to follow
This bitcode should guarantee that he data taken at the second
the pps goes high is the first sample in a vdif packet, the vdif
header for that packet has the new seconds since the epoch, the PSN
updates and the data frame number resets to 0.

This is untested at the moment.