[r2dbe_vdif_lib, r2dbe_top] r2dbe_top is now complete (previously erroneously incomplete) and all nomenclature changed to be compatible with code
The previous 2 commits where I had been re-working the design to
have a more clear hold of the timing and to add the VTP to the
header show that while the r2dbe_vdif_lib file was correct, the
top-level file was incorrectly edited by myself such that the
path for IF0 (includes block vdif_0) was correctly adjusted each
time, but IF1 (includes block vdif_1) remained unchanged.

This commit fixes that error with the r2dbe_top file, and fixes
some minor block name changes that had occurred making the older
software incompatible.  Now VDIF_test.py software used to start
the bitcode works as with previous bitcodes.

At the time of this commit, the receiving software is not compatible
with the addition of the VTP, and Rurik has some unstaged changes.

The bitcode produced from these files is:
... _Aug_27_1442.bof.gz

and is compatible with REV 1 of the R2DBE