[r2dbe_top.slx, r2dbe_rev2.bof.gz] recompile of boffile to fix versionitis
This top file and bitcode should function identically to the previous
bitcode, although there binaries may differ.

When exploring updated software and bitcode changes, we updated the
mlib_devel libraries and merged with the ska.  This worked fine
for simulation of new modules to insert in the r2dbe, but broke
some of the existing functionality.  specifically, the code would
now not compile without the Fixed Point toolbox, because ska had
changed all the software registers to new versions.  Thus when
loading the simulink model file, the registers were instantly
replaced with the newer versions and we could not compile.

We rolled back to the version of mlib_devel, and rolled everything
back to the version in the repo and recompiled.  Small changes in
the placement of things in the model file can cause a new random
seed in the simulated annealing process that XPS does to try and
map the low level components to the FPGA, thus the resulting bit-code
differs only in the low-level implementation, but behaviorally is
identical.

For reference, here is the version of mlib devel, and you can note that
stamped into the bitcode itself is the hash for the mlib devel ver-
sion, which is how we were able to determine exactly which version
of the tools were used to produce it.  Its the rcs block, and I
believe the function is get_rcs() in corr.

     commit a288ff331e933f312ff1c73956b1096ce323a772
     Author: Rurik A Primiani <rprimian@cfa.harvard.edu>
     Date:   Mon Jul 21 21:31:33 2014 -0400

         [adc5g] The area group constraints now check whether
                  it's ROACH2 rev1 or rev